\name{TableToLongForm-package}
\alias{TableToLongForm-package}
\docType{package}
\title{
  Convert a Table to a LongForm data.frame
}
\description{
  TableToLongForm automatically converts hierarchical Tables intended
  for a human reader into a simple LongForm Dataframe that is machine
  readable.
}
\details{
\tabular{ll}{
Package: \tab TableToLongForm\cr
Type: \tab Package\cr
Version: \tab 1.3.1\cr
Date: \tab 2014-08-01\cr
License: \tab GPL-3\cr
}
Call TableToLongForm() on a Table to automatically convert it to a
LongForm data.frame.

Examples of Tables that can be converted are found in data(TCData).

For more details on what TableToLongForm does and what sorts of Tables
it can convert, refer to the website:
\url{https://www.stat.auckland.ac.nz/~joh024/Research/TableToLongForm/}

Available help:
help(TableToLongForm)
help(TCData)
}
\author{
Jimmy Oh

Maintainer: Jimmy Oh <joh024@aucklanduni.ac.nz>
}
\keyword{ package }
