\name{TTLFaliasList}
\alias{TTLFaliasList}
\title{
  List registered Modules for TableToLongForm
}
\description{
  TableToLongForm is partially modular and can be extended in some ways
  with external modules. This function is used to list currently
  registered modules.
}
\details{
  For more details on modules, refer to the ``Working with Modules''
  document on the website:
  \url{https://www.stat.auckland.ac.nz/~joh024/Research/TableToLongForm/}
}